Add the ability to remove an entry from the aliases or related fields.
Introduces the concept of "filters", currently only including one that
removes an entry from the "aliases" or "related" field based on a pattern.

This is a minimal implementation that can be expanded in the future to
include other types of filters (e.g. replace).

For now this change is to support #667

Signed-off-by: Caleb Brown <calebbrown@google.com>