Remove the overly broad match from GHSA to limit scanner noise.
The GitHub repo github.com/primefaces/primereact-sass-theme uses a
package.json file for this name squatting attack and has a version 10.8.5.

The overly broad version range from GitHub that covers every version means
that this repo is matched and if someone has it repo checked out it
will be potentially considered malicious.

Signed-off-by: Caleb Brown <calebbrown@google.com>