Remove open ranges from reports introduces in kam193's PyPI collection
Based on the feedback from community, the open ranges were
problematic when the scanning tool did not recognize where the
package come from and applyied PyPI reports to non-PyPI packages.
To help with handling this, reports were propagated with versions
that existed in PyPI, and this commit handles cleaning up final
reports.

Signed-off-by: Kamil Mankowski (kam193) <me@kam193.eu>